+++ src/config/constants.js
@@ -47,6 +47,7 @@ export const ENDPOINT = {
   CREATE_HEADING_URL: '/create-heading',
   CHANGE_DISPLAY_VALUE: '/bibliographic-record/fixed-field-display-value',
   BIBLIOGRAPHIC_RECORD: '/bibliographic-record',
+  AUTHORITY_RECORD: '/authority-record',
   SUGGEST_TAG_INDICATORS: '/filteredTag',
   SUGGEST_TAG_CODE: '/filteredTagsList',
   HEADING_BY_TAG: '/headings-by-tag',